[Nikhil] - Sage - How to Create your own DAO on Celo (#521)
* [finished] DAO tutorial

* updated users

* tags updated

tags updated according to the recommendation

* tag change - user.tsx

* dependency name change

I accidentally put name of the coverage plugin as solidity-coverage plugin as hardhat-coverage for some reason. Plugin is installed while writing the smart contract but in tutorial while some reason I wrote name as hardhat-coverage.

* Updated formatting

missing formatting of headings was updated here. along with that link to openzeppelin added which was missing.

* minor change

Co-authored-by: Decrypted Bytes <hello@decryptedbytes.com>