Update index.md
Deleted Celo Wallet from the list since is not anymore supported and create a lot of problems and confusion in the users that thinks they are using a product from the Celo Foundation/Clabs. 

Changed the Stakewallet data since they had a rebrending some months ago. 

Added Rabby since is a very valid alternative that should defenetely be in this list.