Revert "fix(scripts/copy-generated-docs): adds `git pull`"
This reverts commit 6f5e22ae8597e87702b970d0c3f9e28d4e063ccd.

Turns out this is a mistake, because we don't want to pull the latest changes from master. Instead, we want to always have the commit of the latest release. Otherwise, the docs might be ahead of released features.

There is a workflow in the celo-org/developer-tooling that opens a PR in the docs repo to update the commit every time a new release is made.