Fix and update CLI docs script (#1312)
* chore(scripts/copy-generated-docs): updates path to CLI docs

* fix(scripts/copy-generated-docs): adds `git pull`

Previously, the script didn't fetch the latest commit from master.
With this change, I'm hoping it will always fetch the latest commit.

* docs(cli): updates files to latest version

Bumps submodule and files to latest commit in celo-org/developer-tooling repo

* docs(cli): updates submodule to latest version

* Revert "fix(scripts/copy-generated-docs): adds `git pull`"

This reverts commit 6f5e22ae8597e87702b970d0c3f9e28d4e063ccd.

Turns out this is a mistake, because we don't want to pull the latest changes from master. Instead, we want to always have the commit of the latest release. Otherwise, the docs might be ahead of released features.

There is a workflow in the celo-org/developer-tooling that opens a PR in the docs repo to update the commit every time a new release is made.