Update paths
Fixes paths such as /path/to/old/datadir/chaindata to be
/path/to/old/datadir/celo/chaindata

Changed the name of environment variable OP_GETH_DIR to OP_GETH_DATADIR
to make it more clear that it is the datadir.

Avoided use of the path `/celo` inside docker containers for the celo
datadir, since it could easily be confused with the child celo dir under the
celo datadir (I used /datadir instead).

Updated the paths for the migration to keep all the files under /migration-files

Removed (I think) unnecessary environment variable setting when running
op-geth in docker '-e GETH_DATADIR=/celo'