Switch operator instructions ordering
This changes the node operator instructions so that op-geth is executed
before op-node, ensuring that op-node won't fail when executed because
it can't find op-geth.

Also removes the step to manually create the jwt token because op-geth
will do this automatically if no token is found at the configured path.